+++ lib/BenchShow/Common.hs
@@ -93,6 +93,8 @@ data GroupStyle =
       Absolute    -- ^ Show absolute field values for all groups
     | Diff        -- ^ Show baseline group values as usual and values for
                   -- the subsequent groups as differences from the baseline
+    | Fraction    -- ^ Show the subsequent group values as a fraction of the
+                  -- baseline group's value.
     | Percent     -- ^ Show baseline group values as 100% and values for
                   -- subsequent groups as a percentage of the baseline
     | PercentDiff -- ^ Baseline group values are shown in regular units.
@@ -405,6 +407,7 @@ getUnitByFieldName fieldName fieldMin =
 fieldUnits :: String -> Double -> GroupStyle -> RelativeUnit
 fieldUnits fieldName fieldMin style =
     case style of
+        Fraction          -> RelativeUnit "x" 1
         Percent           -> RelativeUnit "%" 1
         PercentDiff       -> RelativeUnit "%" 1
         PercentDiffLower  -> RelativeUnit "%" 1
@@ -430,6 +433,9 @@ percentDiffHigher v1 v2 = ((v2 - v1) * 100) / max v1 v2
 percent :: (Fractional a, Num a) => a -> a -> a
 percent v1 v2 = (v2 * 100) / v1
 
+fraction :: (Fractional a, Num a) => a -> a -> a
+fraction v1 v2 = v2 / v1
+
 cmpTransformColumns :: ReportType
                     -> GroupStyle
                     -> Estimator
@@ -451,7 +457,17 @@ cmpTransformColumns rtype style estimator diffStrategy cols =
                 colTransform col = zipWith (mkMinDiff diff) firstCol col
             in map colTransform (tail cols)
 
-        percentDiffWith f =
+        diffWith f =
+                case diffStrategy of
+                    MinEstimator ->
+                        let (ests, vals) = unzip $ map unzip (cmpMinWith f)
+                        in ( Just $ map (const estimator) (head cols) : ests
+                           , head columns : vals
+                           )
+                    SingleEstimator ->
+                        (Nothing, head columns : cmpWith f)
+
+        relativeDiffWith baseVal f =
             -- In a comparative graphical chart we cannot show the absolute
             -- values in the baseline column as the units won't match for
             -- the baseline and the diff clusters.
@@ -461,7 +477,7 @@ cmpTransformColumns rtype style estimator diffStrategy cols =
                         GraphicalChart | length columns == 1 ->
                             head columns
                         GraphicalChart ->
-                            map (\(n,_) -> (n,100)) (head columns)
+                            map (\(n,_) -> (n,baseVal)) (head columns)
             in case diffStrategy of
                 MinEstimator ->
                     let (ests, vals) = unzip $ map unzip (cmpMinWith f)
@@ -471,20 +487,13 @@ cmpTransformColumns rtype style estimator diffStrategy cols =
                 SingleEstimator ->
                    (Nothing, baseCol : cmpWith f)
     in case style of
-            Absolute    -> (Nothing, columns)
-            Percent     -> (Nothing, cmpWith percent)
-            Diff        ->
-                case diffStrategy of
-                    MinEstimator ->
-                        let (ests, vals) = unzip $ map unzip (cmpMinWith absoluteDiff)
-                        in ( Just $ map (const estimator) (head cols) : ests
-                           , head columns : vals
-                           )
-                    SingleEstimator ->
-                        (Nothing, head columns : cmpWith absoluteDiff)
-            PercentDiff       -> percentDiffWith percentDiff
-            PercentDiffLower  -> percentDiffWith percentDiffLower
-            PercentDiffHigher -> percentDiffWith percentDiffHigher
+            Absolute          -> diffWith (\_ x -> x)
+            Fraction          -> relativeDiffWith 1 fraction
+            Diff              -> diffWith absoluteDiff
+            Percent           -> relativeDiffWith 100 percent
+            PercentDiff       -> relativeDiffWith 0 percentDiff
+            PercentDiffLower  -> relativeDiffWith 0 percentDiffLower
+            PercentDiffHigher -> relativeDiffWith 0 percentDiffHigher
     where
         verify a b = if a then b else error "bug: benchmark names mismatch"
         transformVals = map (map (second (getAnalyzedValue estimator)))
@@ -509,18 +518,20 @@ cmpTransformColumns rtype style estimator diffStrategy cols =
 transformColumnNames :: GroupStyle -> [ReportColumn] -> [ReportColumn]
 transformColumnNames _ [] = []
 transformColumnNames style columns@(h:t) =
-    let withDiff = colSuffix baseName h : map (colSuffix diffName) t
+    let withDiff name = colSuffix baseName h : map (colSuffix name) t
     in case style of
-            Diff              | length columns > 1 -> withDiff
-            PercentDiff       | length columns > 1 -> withDiff
-            PercentDiffLower  | length columns > 1 -> withDiff
-            PercentDiffHigher | length columns > 1 -> withDiff
+            Diff              | length columns > 1 -> withDiff diffName
+            Fraction          | length columns > 1 -> withDiff fracName
+            PercentDiff       | length columns > 1 -> withDiff diffName
+            PercentDiffLower  | length columns > 1 -> withDiff diffName
+            PercentDiffHigher | length columns > 1 -> withDiff diffName
             _           -> columns
 
     where
     colSuffix xl col = col { colName = xl (colName col) }
     baseName        = (++ "(base)")
     diffName        = (++ "(-base)")
+    fracName        = (++ "(/base)")
 
 -- Represents the data for a single benchmark run
 data GroupMatrix = GroupMatrix
@@ -1129,10 +1140,7 @@ prepareGroupsReport cfg@Config{..} style outfile rtype runs field matrices =
         mkColUnits :: [RelativeUnit]
         mkColUnits =
             let cols =
-                    if style == Diff
-                        || style == PercentDiff
-                        || style == PercentDiffLower
-                        || style == PercentDiffHigher
+                    if style /= Absolute
                     -- if we consider diff values as well here then the
                     -- units will change to potentially very small.
                     then [head sortedCols]
@@ -1143,20 +1151,18 @@ prepareGroupsReport cfg@Config{..} style outfile rtype runs field matrices =
                         punit = fieldUnits field 1 style -- % unit
                     in unit : replicate (length matrices - 1) punit
                 mkPercentColUnitGraph = [fieldUnits field minVal Absolute]
+                mkAbsoluteUnit =
+                     let unit = fieldUnits field minVal style
+                     in replicate (length matrices) unit
             in case (rtype, style) of
                 -- In case of percentDiff in TextReport we use absolute
                 -- values in the baseline column, so the unit is different.
-                (TextReport, PercentDiff)       -> mkPercentColUnitText
-                (TextReport, PercentDiffLower)  -> mkPercentColUnitText
-                (TextReport, PercentDiffHigher) -> mkPercentColUnitText
-                (GraphicalChart, PercentDiff) | length matrices == 1 ->
+                (_, Absolute) -> mkAbsoluteUnit
+                (_, Diff)     -> mkAbsoluteUnit
+                (TextReport, _)  -> mkPercentColUnitText
+                (GraphicalChart, _) | length matrices == 1 ->
                     mkPercentColUnitGraph
-                (GraphicalChart, PercentDiffLower)  | length matrices == 1 ->
-                    mkPercentColUnitGraph
-                (GraphicalChart, PercentDiffHigher) | length matrices == 1 ->
-                    mkPercentColUnitGraph
-                _ -> let unit = fieldUnits field minVal style
-                     in replicate (length matrices) unit
+                (GraphicalChart, _) -> mkAbsoluteUnit
 
         mkColValues :: [[Double]]
         mkColValues =